Add files via upload
Hello there, this is a project that anyone can use in their work. The project name Wordcounter that counts the number of words and characters and it also track the time in how much minute a person can read that summary . And it has some buttons that converts its states.